iroh_quinn

Struct Path

pub struct Path { /* private fields */ }

An open (Multi)Path

Implementations

impl Path

pub fn id(&self) -> PathId

The PathId of this path.

pub fn status(&self) -> Result<PathStatus, ClosedPath>

The current local [PathStatus] of this path.

pub fn set_status(&self, status: PathStatus) -> Result<(), SetPathStatusError>

Sets the [PathStatus] of this path.

pub fn close(&self) -> Result<ClosePath, ClosePathError>

Closes this path

The future will resolve when all the path state is dropped. This only happens after the remote has confirmed the path as closed and after an additional timeout to give any in-flight packets the time to arrive.

pub fn set_max_idle_timeout( &self, timeout: Option<Duration>, ) -> Result<Option<Duration>, ClosedPath>

Sets the keep_alive_interval for a specific path

See TransportConfig::default_path_keep_alive_interval for details.

Returns the previous value of the setting.

pub fn set_keep_alive_interval( &self, interval: Option<Duration>, ) -> Result<Option<Duration>, ClosedPath>

Sets the keep_alive_interval for a specific path

See TransportConfig::default_path_keep_alive_interval for details.

Returns the previous value of the setting.

pub fn observed_external_addr(&self) -> Result<AddressDiscovery, ClosedPath>

Track changes on our external address as reported by the peer.

If the address-discovery extension is not negotiated, the stream will never return.

pub fn remote_address(&self) -> Result<SocketAddr, ClosedPath>

The peer’s UDP address for this path.

pub fn ping(&self) -> Result<(), ClosedPath>

Ping the remote endpoint over this path.