iroh_quinn

Trait NetworkChangeHint

pub trait NetworkChangeHint: Debug + 'static {
    // Required method
    fn is_path_recoverable(
        &self,
        path_id: PathId,
        network_path: FourTuple,
    ) -> bool;
}

Hints when the caller identifies a network change.

Required Methods

fn is_path_recoverable(&self, path_id: PathId, network_path: FourTuple) -> bool

Inform the connection if a path may recover after a network change.

After network changes, paths may not be recoverable. In this case, waiting for the path to become idle may take longer than what is desirable. If Self::is_path_recoverable returns false, a multipath-enabled, client-side connection will establish a new path to the same remote, closing the current one, instead of migrating the path.

Paths that are deemed recoverable will simply be sent a PING for a liveness check.