iroh_quinn

Trait UdpSender

pub trait UdpSender:
    Send
    + Sync
    + Debug
    + 'static {
    // Required methods
    fn poll_send(
        self: Pin<&mut Self>,
        transmit: &Transmit<'_>,
        cx: &mut Context<'_>,
    ) -> Poll<Result<()>>;
    fn try_send(self: Pin<&mut Self>, transmit: &Transmit<'_>) -> Result<()>;

    // Provided method
    fn max_transmit_segments(&self) -> usize { ... }
}

An object for asynchronously writing to an associated AsyncUdpSocket.

Any number of UdpSenders may exist for a single AsyncUdpSocket. Each UdpSender is responsible for notifying at most one task for send readiness.

Required Methods

fn poll_send( self: Pin<&mut Self>, transmit: &Transmit<'_>, cx: &mut Context<'_>, ) -> Poll<Result<()>>

Send a UDP datagram, or register to be woken if sending may succeed in the future.

Usually implementations of this will poll the socket for writability before trying to write to them, and retry both if writing fails.

Quinn will create multiple UdpSenders, one for each task it’s using it from. Thus it’s important to poll the underlying socket in a way that doesn’t overwrite wakers.

A single UdpSender will be re-used, even if poll_send returns Poll::Ready once, unlike Future::poll, so calling it again after readiness should not panic.

fn try_send(self: Pin<&mut Self>, transmit: &Transmit<'_>) -> Result<()>

Try to send a UDP datagram, if the socket happens to be write-ready.

This may fail with io::ErrorKind::WouldBlock, if the socket is currently full.

The quinn endpoint uses this function when sending

A version negotiation response due to an unknown version

A CLOSE due to a malformed or unwanted connection attempt

A stateless reset due to an unrecognized connection

A Retry packet due to a connection attempt when use_retry is set

If sending in these cases fails, a well-behaved peer will re-try. Thus it’s fine if we drop datagrams sometimes with this function.

Provided Methods

fn max_transmit_segments(&self) -> usize

Maximum number of datagrams that a Transmit may encode.